+++ src/main/java/com/triplan/planner/calendar/dto/DateDTO.java
@@ -1,21 +1,16 @@
 package com.triplan.planner.calendar.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import java.util.Date;
 
 @Data
+@AllArgsConstructor
 public class DateDTO {
 
     private Date startDay;
     private Date endDay;
     private String memberId;
     private String title;
-
-    public DateDTO(Date startDay, Date endDay, String memberId, String title) {
-        this.startDay = startDay;
-        this.endDay = endDay;
-        this.memberId = memberId;
-        this.title = title;
-    }
 }
