+++ src/main/java/com/triplan/planner/calendar/dto/ModifyForm.java
@@ -1,19 +1,15 @@
 package com.triplan.planner.calendar.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import java.util.Date;
 
 @Data
+@AllArgsConstructor
 public class ModifyForm {
 
     private long scheduleNo;
     private Date startDay;
     private Date endDay;
-
-    public ModifyForm(long scheduleNo, Date startDay, Date endDay) {
-        this.scheduleNo = scheduleNo;
-        this.startDay = startDay;
-        this.endDay = endDay;
-    }
 }
