+++ src/main/java/com/triplan/planner/calendar/repository/CalendarRepository.java
@@ -2,25 +2,10 @@ package com.triplan.planner.calendar.repository;
 
 import com.triplan.planner.calendar.dto.DateDTO;
 import com.triplan.planner.calendar.dto.ModifyForm;
-import com.triplan.planner.plan.repository.PlanMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public class CalendarRepository {
+public interface CalendarRepository {
 
-    @Autowired
-    CalendarMapper calendarMapper;
-    @Autowired
-    PlanMapper planMapper;
+    public long saveDate(DateDTO dateDTO);
 
-    public long saveDate(DateDTO dateDTO){
-        calendarMapper.saveDate(dateDTO);
-        return planMapper.getLastScheduleNo();
-    }
-
-    public void modifyDay(ModifyForm modifyForm, long diffmodifyDays) {
-        calendarMapper.deleteOverSchedule(modifyForm.getScheduleNo(), diffmodifyDays);
-        calendarMapper.updateDayByNo(modifyForm);
-    }
+    public void modifyDay(ModifyForm modifyForm, long diffmodifyDays);
 }
