+++ src/main/resources/mapper/tlogMapper.xml
@@ -4,7 +4,7 @@
 
 <mapper namespace="com.triplan.planner.tlog.repository.TlogMapper">
 
-    <insert id="saveTlog" parameterType="Tlog" useGeneratedKeys="true" keyProperty="tlogNo">
+    <insert id="saveTlog" parameterType="com.triplan.planner.tlog.domain.Tlog" useGeneratedKeys="true" keyProperty="tlogNo">
         INSERT INTO TLOG(TLOG_TITLE, TLOG_CONTENT, WRITE_TIME, MEMBER_ID, SCHEDULE_NO)
         VALUES(#{tlogTitle}, #{tlogContent}, now(), #{memberId}, #{scheduleNo});
     </insert>
@@ -17,7 +17,7 @@
         </foreach>
     </insert>
 
-    <select id="getTlogs" parameterType="int" resultType="TlogList">
+    <select id="getTlogs" parameterType="int" resultType="com.triplan.planner.tlog.dto.TlogList">
         SELECT A.TLOG_NO, A.TLOG_TITLE, A.TLOG_CONTENT, A.MEMBER_ID, B.STORE_NAME
         FROM TLOG A JOIN (SELECT STORE_NAME, TLOG_NO
                           FROM (SELECT STORE_NAME, TLOG_NO, ROW_NUMBER() OVER (PARTITION BY TLOG_NO ORDER BY (SELECT NULL)) AS RN
@@ -33,13 +33,13 @@
         FROM TLOG;
     </select>
 
-    <select id="getTlogByNo" parameterType="long" resultType="Tlog">
+    <select id="getTlogByNo" parameterType="long" resultType="com.triplan.planner.tlog.domain.Tlog">
         SELECT *
         FROM TLOG
         WHERE TLOG_NO=#{tlogNo};
     </select>
 
-    <select id="getTlogImageByNo" parameterType="long" resultType="TlogImage">
+    <select id="getTlogImageByNo" parameterType="long" resultType="com.triplan.planner.tlog.domain.TlogImage">
         SELECT *
         FROM TLOGIMAGE
         WHERE TLOG_NO=#{tlogNo};
@@ -50,7 +50,7 @@
         WHERE TLOG_NO=#{tlogNo}
     </delete>
 
-    <update id="updateTlog" parameterType="Tlog">
+    <update id="updateTlog" parameterType="com.triplan.planner.tlog.domain.Tlog">
         UPDATE TLOG
         SET TLOG_TITLE=#{tlogTitle}, TLOG_CONTENT=#{tlogContent}, SCHEDULE_NO=#{scheduleNo}
         WHERE TLOG_NO=#{tlogNo};
